+++ docs/doctest_header.py
@@ -25,9 +25,11 @@ from dffml.df.base import *
 from dffml.df.types import *
 from dffml.util.net import *
 from dffml.df.memory import *
+from dffml.model.slr import *
 from dffml_model_scikit import *
 from dffml.operation.io import *
 from dffml.source.memory import *
+from dffml.operation.model import *
 from dffml.operation.output import *
 from dffml.operation.dataflow import *
 from dffml.operation.preprocess import *