+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Doctestable examples to `operation/mapping.py`.
 ## [0.3.7] - 2020-04-14
 ### Added
 - IO operations demo and `literal_eval` operation.