+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Doctestable examples to `operation/mapping.py`
+- load function in high_level API
 
 ## [0.3.7] - 2020-04-14
 ### Added